+++ runtime/dancebox/tests/common/xcm/mod.rs
@@ -16,6 +16,7 @@
 
 mod constants;
 mod core_buyer;
+mod core_buyer_common;
 mod delivery_fees;
 mod force_core_buyer;
 mod foreign_signed_based_sovereign;
